requirements.txt
annotated-types==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.2.7
db-dtypes==1.2.0
exceptiongroup==1.2.2
Flask==2.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==0.2.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==2.6.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
requests==2.32.3
rsa==4.9
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
python-dotenv==0.19.0